@@ -2,36 +2,36 @@
   "name": "quick-look-table",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "type": "module",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development vite",
+    "dev": "vite dev",
     "build": "vite build",
-    "test": "jest",
-    "test:watch": "jest --watchAll",
-    "serve": "vite preview"
+    "preview": "vite preview",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^1.0.1",
-    "@testing-library/svelte": "^3.0.3",
-    "@types/jest": "^26.0.21",
-    "autoprefixer": "^10.2.5",
-    "axios": "^0.27.2",
-    "babel-jest": "^26.6.3",
-    "cross-env": "^7.0.3",
-    "cssnano": "^5.1.12",
-    "jest": "^26.6.3",
-    "postcss": "^8.4.16",
-    "postcss-load-config": "^4.0.1",
-    "prettier": "^2.7.1",
-    "prettier-plugin-svelte": "^2.7.0",
-    "svelte": "^3.50.1",
-    "svelte-jester": "^1.3.1",
-    "svelte-preprocess": "^4.10.7",
-    "tailwindcss": "^3.1.8",
-    "ts-jest": "^26.5.4",
-    "typescript": "^4.8.3",
-    "vite": "^3.0.0"
+    "@sveltejs/adapter-static": "3.0.10",
+    "@sveltejs/kit": "2.50.0",
+    "@sveltejs/vite-plugin-svelte": "6.2.4",
+    "autoprefixer": "10.4.23",
+    "axios": "1.13.2",
+    "cssnano": "7.1.2",
+    "dayjs": "1.11.19",
+    "postcss": "8.5.6",
+    "postcss-load-config": "6.0.1",
+    "prettier": "3.8.1",
+    "prettier-plugin-svelte": "3.4.1",
+    "prettier-plugin-tailwindcss": "0.7.2",
+    "svelte": "5.47.1",
+    "svelte-check": "4.3.5",
+    "svelte-preprocess": "6.0.3",
+    "tailwindcss": "3.4.17",
+    "tslib": "2.8.1",
+    "typescript": "5.9.3",
+    "vite": "6.3.4"
   },
   "author": "ascarbek@gmail.com",
-  "license": "MIT"
+  "license": "MIT",
+  "packageManager": "pnpm@10.10.0+sha512.d615db246fe70f25dcfea6d8d73dee782ce23e2245e3c4f6f888249fb568149318637dca73c2c5c8ef2a4ca0d5657fb9567188bfab47f566d1ee6ce987815c39"
 }
